add: notify frontend on all cache changes
This allows us to react more timely to cache changes. Otherwise,
it frequently happend that we displayed a node as lagging behind
for a while until we fetched data for the new block. We queue very
frequent data fetches up on the frontend, but frequently hitting the
backend shouldn't be a problem, as the backend caches the responses.